+++ test/sleep.js
@@ -10,5 +10,5 @@ module.exports =
    * @ignore
    */
   function sleep(ms) {
-    return new Promise(resolve => setTimeout(resolve, ms))
+    return new Promise((resolve) => setTimeout(resolve, ms))
   }