+++ test/fixtures/output.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { TestDirector } = require('../../')
+const TestDirector = require('../../lib/TestDirector.js')
 
 const tests = new TestDirector()
 tests.add('a', () => {