+++ changelog.md
@@ -7,6 +7,7 @@
 - Updated dependencies.
 - Lint fixes for [`prettier`](https://npm.im/prettier) v2.
 - Ensure GitHub Actions run on pull request.
+- Moved the `simulatePublishedTraces` test helper into its own file.
 - Tidied the position of a JSDoc comment.
 
 ## 3.0.1